[2398] Reduce test run times by preventing migration after each test
This commit leverages Micronaut TestResources to manage the postgres database container for the tests.

This means that the container will exist across all the tests and so we save time waiting for it to start up and migrate each test class.

Locally I have seen the time to run all the tests drop from 4m55s to 2m11s