[2237] Test Containers Fail To Start For Classes Not Using The TestContainersSuite
One option is to have all classes extend the TestContainersSuite, but with this approach running all tests took about 2min 30s. With the changes in this PR it shaves off about 40s leaving us at roughly 1min 50s. With that in mind this seems to be the way to go.