Reinstate RequestLogging for Local ENV
In Micronaut 4, the request attributes changed for routing.
This means that our request logger broke as it was looking for routeInfo on the "micronaut.http.route" attribute which no-longer exists.

This PR switches to use the new attribute to re-enable the logging, and introduces an order so we can capture security information.

After discussion, this logger is only enabled when the LOCAL environment is enabled.