[sonar] remove useless Mockito eq usage
Mockito provides argument matchers for flexibly stubbing or verifying method calls.

`Mockito.verify()`, `Mockito.when()`, `Stubber.when()` and `BDDMockito.given()` each have overloads with and without argument matchers.

However, the default matching behavior (i.e. without argument matchers) uses `equals()`.
If only the matcher `org.mockito.ArgumentMatchers.eq()` is used, the call is equivalent to the call without matchers, i.e. the `eq()` is not necessary and can be omitted.
The resulting code is shorter and easier to read.