build: enable download-sources in IntelliJ Idea
When you view a de-compiled third-party class and click 'Download sources' in IntelliJ, it fails with an error about being unable to find the 'web-ui' project.

This is because of an empty settings.gradle file in the server directory.

This PR removes that file, and fixes source downloading in the IDE