Fix GoogleServiceConfiguration field names
This fixes a couple of issues:

1. We had non-idiomatic field names in `GoogleServiceConfiguration`
2. We were building JSON via toString()

To fix this, we use standard Java naming in the config object (Micronaut handles the dash or underscore replacement from config)

And we use `@JsonProperty` and Micronaut's `JsonMapper` to perform the serialization.

Removed the test for getters/setters working and replaced with a test that verifies config can be loaded, and converted to a json string and parsed.