docs: Update troubleshooting to show db rebuild
This commit updates the troubleshooting page to show how to rebuild the database with `podman compose`.

I looked in to adding tabbed controls to the page to show the analog docker commands.
But it was an awful lot of code for a simple search and replace of `podman` for `docker`.

So I added a section to say you can just switch the word.