Add an object to contain GitHub configuration
We also move the base URL into configuration so that we can test the client against a mock server.
Added a test which mocks up a GitHub server, and tests agains JSON from the documentation to make sure it's all working.

I'm not sure this is used however, as the GithubClient used to point to `repos/oci-labs/check-ins/issues` and I cannot find a check-ins project under the `oci-labs` owner...🤔